+++ R/read_tracts.R
@@ -61,7 +61,6 @@ read_tracts <- function(year,
   ### check inputs
   if (missing(year) || is.null(year)) { error_year_not_declared() }
   checkmate::assert_numeric(year, any.missing = FALSE)
-  checkmate::assert_string(dataset, null.ok = FALSE)
   checkmate::assert_logical(as_data_frame)
   checkmate::assert_logical(showProgress)
   checkmate::assert_logical(cache)
@@ -85,6 +84,15 @@ read_tracts <- function(year,
   data_sets_2022 <- c("Basico", "Domicilio", "Pessoas", "ResponsavelRenda",
                       "Indigenas", "Quilombolas", "Entorno", "Obitos",
                       "Preliminares")
+
+  # data sets available for the requested year
+  data_sets_year <- switch(as.character(year),
+                           '2000' = data_sets_2000,
+                           '2010' = data_sets_2010,
+                           '2022' = data_sets_2022)
+
+  if (missing(dataset) || is.null(dataset)) { error_arg_not_declared('dataset', data_sets_year) }
+  checkmate::assert_string(dataset, null.ok = FALSE)
 
   # check requested data set
   dataset <- tolower(dataset)